All Game Attendance Data

This data set contains information about the attendance of all games in a certain season or series. It includes columns such as game date, attendance numbers, and other related statistics.

Usage

data(All_Game_Attendance)

Format

A data frame with 6214 rows and 13 variables:

...1

Number of games played

Year

Integer. The year the teams played.

Day of Week

Factor. The day of the week the teams played.

Month

Factor. The month the teams played.

Day

Integer. Day of the week the teams played.

Time (ET)

The time the game was played at.

Game Type

Factor. The game is classified either as regular season, Seminfinals, Finals, Eastern Conference Semifinals etc.

Home Team

Fcator. Name of the home team.

Away Team

Factor. Name of the away team.

Arena

Factor. Name of the arena the game was played at.

City

Factor. Name of the city the game was played at.

State

Factor. Name of the State the game was played at.

Attendance

Integer. Number of people who attended the game.

Source

https://www.acrossthetimeline.com/wnba/data.html#