Function shows career PPG of top WNBA scorers

This function scrapes career points per game (PPG) data for specified WNBA players from Basketball Reference and returns an interactive plot showing points scored over the years.

Usage

career_ppg(players = NULL)

Source

https://www.basketball-reference.com/wnba/players/

Arguments

players

the first and last name of WNBA player(s)

Value

An interactive plot showing the career points scored by WNBA players over the years.

Details

This function fetches career points (PTS) and year (Year) data for top WNBA players, converts the data to numeric, and visualizes it using an interactive plot.

Author

Jane Loo

Examples

career_ppg()
#> Warning: There were 2 warnings in `dplyr::mutate()`.
#> The first warning was:
#> ℹ In argument: `Year = ifelse(grepl("^[0-9]{4}$", Year), as.numeric(Year),
#>   NA)`.
#> Caused by warning in `ifelse()`:
#> ! NAs introduced by coercion
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 2 warnings in `dplyr::mutate()`.
#> The first warning was:
#> ℹ In argument: `Year = ifelse(grepl("^[0-9]{4}$", Year), as.numeric(Year),
#>   NA)`.
#> Caused by warning in `ifelse()`:
#> ! NAs introduced by coercion
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There was 1 warning in `dplyr::mutate()`.
#> ℹ In argument: `Year = ifelse(grepl("^[0-9]{4}$", Year), as.numeric(Year),
#>   NA)`.
#> Caused by warning in `ifelse()`:
#> ! NAs introduced by coercion
#> Warning: There were 2 warnings in `dplyr::mutate()`.
#> The first warning was:
#> ℹ In argument: `Year = ifelse(grepl("^[0-9]{4}$", Year), as.numeric(Year),
#>   NA)`.
#> Caused by warning in `ifelse()`:
#> ! NAs introduced by coercion
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There was 1 warning in `dplyr::mutate()`.
#> ℹ In argument: `Year = ifelse(grepl("^[0-9]{4}$", Year), as.numeric(Year),
#>   NA)`.
#> Caused by warning in `ifelse()`:
#> ! NAs introduced by coercion
#> Warning: There was 1 warning in `dplyr::mutate()`.
#> ℹ In argument: `Year = ifelse(grepl("^[0-9]{4}$", Year), as.numeric(Year),
#>   NA)`.
#> Caused by warning in `ifelse()`:
#> ! NAs introduced by coercion
#> Warning: There was 1 warning in `dplyr::mutate()`.
#> ℹ In argument: `Year = ifelse(grepl("^[0-9]{4}$", Year), as.numeric(Year),
#>   NA)`.
#> Caused by warning in `ifelse()`:
#> ! NAs introduced by coercion
#> Warning: There was 1 warning in `dplyr::mutate()`.
#> ℹ In argument: `Year = ifelse(grepl("^[0-9]{4}$", Year), as.numeric(Year),
#>   NA)`.
#> Caused by warning in `ifelse()`:
#> ! NAs introduced by coercion
#> Warning: There was 1 warning in `dplyr::mutate()`.
#> ℹ In argument: `Year = ifelse(grepl("^[0-9]{4}$", Year), as.numeric(Year),
#>   NA)`.
#> Caused by warning in `ifelse()`:
#> ! NAs introduced by coercion
#> Warning: There was 1 warning in `dplyr::mutate()`.
#> ℹ In argument: `Year = ifelse(grepl("^[0-9]{4}$", Year), as.numeric(Year),
#>   NA)`.
#> Caused by warning in `ifelse()`:
#> ! NAs introduced by coercion

career_ppg("Diana Taurasi")
#> Warning: There were 2 warnings in `dplyr::mutate()`.
#> The first warning was:
#> ℹ In argument: `Year = ifelse(grepl("^[0-9]{4}$", Year), as.numeric(Year),
#>   NA)`.
#> Caused by warning in `ifelse()`:
#> ! NAs introduced by coercion
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.